+++ assets/js/script.js
@@ -2,18 +2,6 @@
 The general script for the all the pages.
 Used init bootstrap tooltips, popovers and modals throughout the site.
 */
-import { setDifficulty, setCurrentQuestion, setProgress } from "./game-storage.js";
-
-// Set the difficulty level when the play button is clicked.
-const playButton = document.getElementById("play-btn");
-playButton.addEventListener("click", () => {
-  const difficulty = playButton.dataset.difficulty;
-
-  setDifficulty(difficulty);
-  setCurrentQuestion();
-  setProgress(0);
-  window.location.href = "game.html";
-});
 
 const rulesButton = document.getElementById("rulesButton");
 const rulesModal = document.getElementById("rulesModal");
