+++ index.html
@@ -143,7 +143,8 @@
 
   <!-- Custom JS -->
   <script type="module" src="assets/js/game-storage.js"></script>
-  <script type="module" src="assets/js/script.js"></script>
+  <script type="module" src="assets/js/home.js"></script>
+  <script src="assets/js/script.js"></script>
 </body>
 
 </html>
